Fixes and Borderless modes
Majority was just fixing some errors on my part.

But I also:
implemented switching between borderless modes

also added an option to disable the config copying function in sequential launches. it will still do it on the initial launch.

Not ready for a exe compile of this version just yet. got some messages blocks to edit before that.