V.0.6.3.1
Minor update to further protect against what occurred that required the emergency fix. 

Long story short: when after the launcher was launched with a accepted parameter, the launcher tried to relaunch itself to fix values. however the relaunch section was copied from towards the top of the script where it relaunches the launcher asking for admin privileges. 

WHICH meant it still had the %1% launch parameter. So it was jumping to the game launch section over and over and over and over and over.